feat: Refactor calendar screen
