+++ lib/presentation/widgets/screens/calendar/builders/calendar_builders_helper.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:table_calendar/table_calendar.dart';
-import 'package:dienstplan/presentation/widgets/screens/calendar/calendar_date_selector/animated_calendar_day.dart';
+import 'package:dienstplan/presentation/widgets/screens/calendar/date_selector/animated_calendar_day.dart';
 import 'package:dienstplan/presentation/controllers/schedule_controller.dart';
 
 class CalendarBuildersHelper {
