+++ lib/presentation/widgets/screens/calendar/utils/schedule_list_animation_mixin.dart
